+++ bin/ruk-standalone.ts
@@ -1,5 +1,3 @@
-#!/usr/bin/env node
-
 import { start } from "../src/entrypoint.js";
 
-start("package");
+start("standalone");